get last known location feature implemented

+++ app/src/main/java/eu/vmpay/weathermate/utils/rest/Main.java
@@ -12,10 +12,10 @@ public class Main
 	private Double temp;
 	@SerializedName("pressure")
 	@Expose
-	private Integer pressure;
+	private Double pressure;
 	@SerializedName("humidity")
 	@Expose
-	private Integer humidity;
+	private Double humidity;
 	@SerializedName("temp_min")
 	@Expose
 	private Double tempMin;
@@ -33,22 +33,22 @@ public class Main
 		this.temp = temp;
 	}
 
-	public Integer getPressure()
+	public Double getPressure()
 	{
 		return pressure;
 	}
 
-	public void setPressure(Integer pressure)
+	public void setPressure(Double pressure)
 	{
 		this.pressure = pressure;
 	}
 
-	public Integer getHumidity()
+	public Double getHumidity()
 	{
 		return humidity;
 	}
 
-	public void setHumidity(Integer humidity)
+	public void setHumidity(Double humidity)
 	{
 		this.humidity = humidity;
 	}
